# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API total backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Registrar un nuevo usuario",
        "tags": ["Autenticación"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "rol": { "type": "string", "example": "admin" },
                  "username": { "type": "string", "example": "usuario1" },
                  "correo": { "type": "string", "example": "usuario@correo.com" },
                  "password": { "type": "string", "example": "123456" }
                },
                "required": ["rol", "username", "correo", "password"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuario registrado exitosamente" },
          "400": { "description": "Errores de validación" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Iniciar sesión",
        "tags": ["Autenticación"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "usuario1" },
                  "password": { "type": "string", "example": "123456" }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login exitoso" },
          "400": { "description": "Errores de validación" }
        }
      }
    },
    "/auth/forgot-password": {
      "post": {
        "summary": "Solicitar restablecimiento de contraseña",
        "tags": ["Autenticación"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "correo": { "type": "string", "example": "usuario@correo.com" }
                },
                "required": ["correo"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Correo de recuperación enviado" },
          "400": { "description": "Errores de validación" }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Restablecer contraseña",
        "tags": ["Autenticación"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string", "example": "token123" },
                  "newPassword": { "type": "string", "example": "nuevacontraseña123" }
                },
                "required": ["token", "newPassword"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Contraseña restablecida exitosamente" },
          "400": { "description": "Errores de validación" }
        }
      }
    },
...
  }
}